+++ GameLib.Plugin/GameLib.Plugin.Rockstar/Model/RockstarGame.cs
@@ -8,25 +8,14 @@ public class RockstarGame : IGame
 {
     #region Interface implementations
     public string Id { get; internal set; } = string.Empty;
-
     public Guid LauncherId { get; internal set; } = Guid.Empty;
-
     public string Name { get; internal set; } = string.Empty;
-
     public string InstallDir { get; internal set; } = string.Empty;
-
-    public string ExecutablePath { get; internal set; } = string.Empty;
-
     public string Executable { get; internal set; } = string.Empty;
-
-    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
-
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(Executable);
     public string WorkingDir { get; internal set; } = string.Empty;
-
     public string LaunchString { get; internal set; } = string.Empty;
-
     public DateTime InstallDate { get; internal set; } = DateTime.MinValue;
-
     public bool IsRunning => ProcessUtil.IsProcessRunning(Executable);
     #endregion